+++ packages/ocpp-ws-io/src/index.ts
@@ -26,6 +26,7 @@ export {
   RPCMessageTypeNotSupportedError,
   RPCNotImplementedError,
   RPCNotSupportedError,
+  RPCOccurenceConstraintViolationError,
   RPCOccurrenceConstraintViolationError,
   RPCPropertyConstraintViolationError,
   RPCProtocolError,